Update dependencies docs
I've also added a test for the example in here.
